requirements.txt
discord.py==2.0.0
python-dotenv==0.20.0
yt_dlp==2022.8.19
